+++ src/components/player-rank/player-rank.tsx
@@ -5,7 +5,7 @@ import TableContainer from '@/components/common/table/table-container';
 import PlayerRankTable from '@/components/player-rank/player-rank-table';
 import useLeagueName from '@/hooks/useLeagueName';
 import useLeagueSeason from '@/hooks/use-league-season';
-import { leagueNameMapper } from '@/utils/leagueMapper';
+import { leagueNameMapper } from '@/utils/league-mapper';
 
 export default function PlayerRank() {
   const leagueName = leagueNameMapper[useLeagueName() as keyof typeof leagueNameMapper];
