+++ src/components/rank/rank.tsx
@@ -5,7 +5,7 @@ import RankTable from '@/components/rank/rank-table';
 import { useGetTeamRank } from '@/api/league';
 import { LoadingBox } from '@/components/common/loading-spinner';
 import useLeagueName from '@/hooks/useLeagueName';
-import { leagueIdMapper } from '@/utils/leagueMapper';
+import { leagueIdMapper } from '@/utils/league-mapper';
 
 export default function Rank() {
   // Todo: 현재 2023 팀랭크만 구현되어 2023으로 임시 설정
@@ -13,7 +13,7 @@ export default function Rank() {
   // const fullSeason = '20' + season;
   const fullSeason = 2023;
   const leagueId = leagueIdMapper[useLeagueName() as keyof typeof leagueIdMapper];
-  const { isPending, data, error } = useGetTeamRank({ season: +fullSeason, leagueId });
+  const { isPending, data, error } = useGetTeamRank({ season: fullSeason.toString(), leagueId });
 
   return (
     <>
